moved OrdersPanel to components folder
